requirements.txt
Django==2.1.7
django-allauth==0.39.1
django-debug-toolbar==1.11.0
django-filter==2.1.0
django-isbn-field==0.5.2
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
psycopg2==2.7.7
pyyaml==3.13
